third_party/WebKit/LayoutTests/fast/dom/Window/window-constructor-presence.html

<!DOCTYPE html>
<body>
<script src="../../../resources/js-test.js"></script>
<script>

description('Tests that the Window constructor is present and that window is an instanceof of it.');

shouldNotBe('typeof Window', '"undefined"');
shouldBeTrue('window instanceof Window');

</script>
</body>
