third_party/WebKit/LayoutTests/fast/dom/Window/window-constructor-presence-expected.txt

Tests that the Window constructor is present and that window is an instanceof of it.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS typeof Window is not "undefined"
PASS window instanceof Window is true
PASS successfullyParsed is true

TEST COMPLETE
